주문기능구현중

+++ src/main/java/springpractice/shoppingmall/DTO/ProductUpdateDto.java
@@ -8,6 +8,5 @@ public class ProductUpdateDto {
     private String name;
     @Min(value = 0 )
     private Long price;
-
-    private int quantity;
+    private Long quantity;
 }
